+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## [0.1.0](https://github.com/webhippie/gettext_i18n_rails_js/releases/tag/v0.1.0) - 2015-02-19
+## [0.1.0](https://github.com/webhippie/gettext_i18n_rails_js/releases/tag/v0.1.0) - 2015-02-24
 
 * Transfer from github.com/nubis/gettext_i18n_rails_js
 * Added TravisCI, Rubocop and Coveralls
 * Updated structure to my opinionated gem style
+* Changed default handlebars function to ```__```
+* Added extended ```gettext_i18n_rails_js.yml```